# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Микросервис для работы с геопространственными данными из OpenStreetMap. Предоставляет API для поиска административных границ, транспортной инфраструктуры, точек интереса (POI), а также векторных тайлов для визуализации на картах.\n\nОсновные возможности:\n- Поиск и обратное геокодирование административных границ\n- Поиск ближайших транспортных станций (метро, автобусы, трамваи)\n- Поиск точек интереса по категориям в радиусе\n- Получение векторных тайлов (MVT/PBF) для всех типов данных\n- Статистика по загруженным данным",
        "title": "Location Microservice API",
        "contact": {
            "name": "API Support",
            "email": "support@location-microservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/debug/enrichment/location": {
            "get": {
                "description": "Тестирует полный процесс обогащения локации. GET версия для быстрого тестирования в браузере.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrichment Debug"
                ],
                "summary": "Обогащение локации (GET дебаг/тест)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Страна (обязательно)",
                        "name": "country",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Регион",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Провинция",
                        "name": "province",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Город",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Район",
                        "name": "district",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Район/квартал",
                        "name": "neighborhood",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Улица",
                        "name": "street",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Номер дома",
                        "name": "house_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Почтовый индекс",
                        "name": "postal_code",
                        "in": "query"
                    },
                    {
                        "maximum": 90,
                        "minimum": -90,
                        "type": "number",
                        "description": "Широта",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "maximum": 180,
                        "minimum": -180,
                        "type": "number",
                        "description": "Долгота",
                        "name": "lon",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обогащённая локация",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Тестирует полный процесс обогащения локации: резолвинг границ, поиск транспорта. Возвращает обогащённые данные в том же формате, что и worker обогащения.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrichment Debug"
                ],
                "summary": "Обогащение локации (дебаг/тест)",
                "parameters": [
                    {
                        "description": "Данные локации для обогащения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обогащённая локация",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/debug/enrichment/location/batch": {
            "post": {
                "description": "Обогащает пачку локаций эффективно (2 параллельных запроса в БД). Локации делятся на 2 группы: с координатами (reverse geocoding) и без (поиск по названиям). Оптимально для обработки 50+ локаций за раз.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrichment Debug"
                ],
                "summary": "Батч обогащение локаций (дебаг/тест)",
                "parameters": [
                    {
                        "description": "Данные локаций для обогащения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обогащённые локации",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationBatchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/debug/enrichment/transport/batch": {
            "post": {
                "description": "Возвращает ближайшие станции общественного транспорта с информацией о линиях для пачки координат одним эффективным запросом. Идеально для обогащения множества объектов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enrichment Debug"
                ],
                "summary": "Batch получение ближайших станций транспорта для нескольких точек",
                "parameters": [
                    {
                        "description": "Batch параметры поиска",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugTransportBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты для каждой точки",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugTransportBatchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_location-microservice_internal_pkg_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_errors.AppError"
                }
            }
        },
        "github_com_location-microservice_internal_pkg_utils.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "time_ms": {
                    "type": "number"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_location-microservice_internal_pkg_utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/github_com_location-microservice_internal_pkg_utils.Meta"
                }
            }
        },
...
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationBatchRequest": {
            "type": "object",
            "required": [
                "locations"
            ],
            "properties": {
                "locations": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.LocationInput"
                    }
                }
            }
        },
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationBatchResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.LocationBatchMeta"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.LocationEnrichmentResult"
                    }
                }
            }
        },
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationRequest": {
            "type": "object",
            "required": [
                "country"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string",
                    "minLength": 2
                },
                "district": {
                    "type": "string"
                },
                "house_number": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "neighborhood": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugLocationResponse": {
            "type": "object",
            "properties": {
                "enriched_location": {
                    "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichedLocationDTO"
                },
                "error": {
                    "type": "string"
                },
                "nearest_transport": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.NearestStationDTO"
                    }
                }
            }
        },
...
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugTransportBatchRequest": {
            "type": "object",
            "required": [
                "points"
            ],
            "properties": {
                "max_distance": {
                    "description": "метры",
                    "type": "number",
                    "maximum": 10000,
                    "minimum": 100
                },
                "points": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.TransportSearchPointDTO"
                    }
                }
            }
        },
        "github_com_location-microservice_internal_usecase_dto.EnrichmentDebugTransportBatchResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.EnrichmentDebugBatchMeta"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_location-microservice_internal_usecase_dto.PointTransportResult"
                    }
                }
            }
        },
...
    }
}